+++ api/v1/class/auth/plugins/mail_password_changed.auth.arbeit.inc.php
@@ -15,7 +15,7 @@ namespace Arbeitszeit{
             if($count == 1){
                 $data = mysqli_fetch_assoc($res);
             } else {
-                Exceptions::error_rep("An error occured while fetching user data from database for user '$username' | SQL-Error: " . mysqli_error($res1));
+                Exceptions::error_rep("An error occured while fetching user data from database for user '$username' | SQL-Error: " . mysqli_error($conn));
                 return [
                     "error" => [
                         "error_code" => 10,